requirements.txt
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# WebSocket support
websockets>=14.0

# HTTP client for Redfish
httpx>=0.28.0
requests>=2.32.0

# Background tasks
apscheduler>=3.10.4

# Template engine
jinja2>=3.1.3

# Utilities
python-dateutil>=2.9.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
